+++ src/FoodStorage/Core.Domain/Domain.Entities/Common/DomainEvents/AddedProductItemDomainEvent.cs
@@ -15,9 +15,9 @@ public class AddedProductItemDomainEvent : BaseDomainEvent
     /// <summary>
     /// Количество продукта
     /// </summary>
-    public int ProductCount { get; }
+    public double ProductCount { get; }
 
-    public AddedProductItemDomainEvent(ProductId productId, int productCount, UserId reducedBy, DateTime reducedAt) : base(reducedBy, reducedAt)
+    public AddedProductItemDomainEvent(ProductId productId, double productCount, UserId reducedBy, DateTime reducedAt) : base(reducedBy, reducedAt)
     {
         ProductId = productId;
         ProductCount = productCount;
